+++ Sources/BartyCrouchTranslator/DeeplApi/Model/DeepLTranslateErrorResponse.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+struct DeepLTranslateErrorResponse: Decodable {
+    let message: String
+}